+++ hc/accounts/tests/test_createsuperuser.py
@@ -19,3 +19,14 @@ class CreateSuperuserTestCase(BaseTestCase):
 
         u = User.objects.get(email="superuser@example.org")
         self.assertTrue(u.is_superuser)
+
+    def test_it_rejects_duplicate_email(self):
+        cmd = Command(stdout=Mock(), stderr=Mock())
+        with patch(cmd.__module__ + ".input") as mock_input:
+            with patch(cmd.__module__ + ".getpass") as mock_getpass:
+                mock_input.side_effect = ["alice@example.org", "alice2@example.org"]
+                mock_getpass.return_value = "hunter2"
+                cmd.handle()
+
+        u = User.objects.get(email="alice2@example.org")
+        self.assertTrue(u.is_superuser)